fix(spec,notebook): os tres bloqueantes derrotavam o proprio proposito
Revisao adversarial do #216: seis achados, tres bloqueantes, nenhum pego
pelos 3.176 testes verdes. Todos medidos rodando o binario.

BLOQUEANTE — o `line` era u64 aqui e u32 no leitor, e o comentario que
eu escrevi jurava que as duas formas batiam byte a byte. `--line
4294967296` gravava um arquivo que o `spec-draft` RECUSA, e o
carregador dele e fail-closed de proposito: a partir dali o canal de
material da unidade morria, porque nao existe porta para remover o item.
O defeito que este comando existe para fechar, causado por ele mesmo.

BLOQUEANTE — um arquivo de material corrompido era descartado em
silencio e sobrescrito. Medido: tres itens gravados, arquivo truncado,
mais um `material-add`, e os tres sumiram — reportado como ok. O leitor
e fail-closed exatamente porque "o material some e ninguem e avisado" e
o defeito; o escritor reintroduzia isso. Agora recusa, nomeia o que
estava em jogo, e nao escreve nada. Arquivo VAZIO nao e corrompido: uma
criacao interrompida deixa zero bytes, e recusar ali travaria a unidade
por nada.

BLOQUEANTE — o portao cortava espacos antes de procurar o marcador e o
`parse_items` nao. Uma linha indentada era vista pelo portao e invisivel
para o caderno; como o `render` reescreve o arquivo a partir dos itens,
o proximo `--add` a APAGAVA e o fechamento passava. Um recuo automatico
de editor bastava. Dois leitores do mesmo arquivo precisam concordar
sobre o que e uma linha.

MAIOR — o aviso novo afirmava que os hooks ainda estavam rodando a
versao antiga, mas o processo que imprime a frase E o binario mais novo.
A frase passa a dizer o que e verdade: o que ficou para tras sao os
comandos, skills e agentes do plugin.

MENORES — a funcao nova entrou entre a documentacao do `run_close_gates`
e a propria funcao, roubando o atributo e a doc dela. E o marcador podia
ser forjado digitando o prefixo no texto: agora a flag e a unica coisa
que o define, e um prefixo digitado e removido, sobrevivendo como texto
em vez de armar um portao.

Provas negativas tomadas nos dois bloqueantes testaveis: desfazendo a
correcao, o teste novo falha.

+++ apps/rt/src/commands/spec/cli.rs
@@ -328,8 +328,11 @@ pub enum SpecCmd {
         #[arg(long)]
         detail: String,
         /// A finding's line number, when the claim is line-precise.
+        ///
+        /// `u32`, the width `spec-draft --material` reads. A wider one here
+        /// would let the door accept a value the draft refuses.
         #[arg(long)]
-        line: Option<u64>,
+        line: Option<u32>,
     },
     /// Deliberately change ONE acceptance criterion after the spec artefacts are
     /// frozen, and prove the replacement still knows how to fail.
